pyproject.toml
[project]
name = "substr8"
version = "1.5.0"
description = "Substr8 Platform CLI - Verifiable AI Infrastructure"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Substr8 Labs", email = "hello@substr8labs.com" }
]
keywords = ["ai", "agents", "verifiable", "memory", "capability", "fdaa", "gam", "acc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
fdaa = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-jaeger>=1.20.0",
]
fdaa-server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "motor>=3.0.0",
]
retrieval = [
    "sentence-transformers>=2.0",
    "numpy>=1.24.0",
]
search = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
]
search-local = [
    "sentence-transformers>=2.0",
    "chromadb>=0.4.0",
]
gam-server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
gam-client = [
    "httpx>=0.24.0",
]
gam-postgres = [
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
]
gam-api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
]
mcp = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
]
full = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "cryptography>=41.0",
    "sentence-transformers>=2.0",
    "chromadb>=0.4.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
substr8 = "substr8.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.venv",
    "/target",
    "/bridge",
    "/docker", 
    "/src",
    "Cargo.toml",
    "Cargo.lock",
    "*.sqlite",
    "__pycache__",
    "*.pyc",
    ".git",
]

[tool.hatch.build.targets.wheel]
packages = ["substr8"]

[tool.ruff]
line-length = 100
target-version = "py310"
